requirements.txt
mysqlclient
django
ipython
djangorestframework
markdown
django-filter

